requirements.txt
torch
numpy
pandas
typing
wandb
transformers
hydra-core==1.2.0